Position dots radially around the node
Created attributes for the features objects **xDot** and **yDot** that compute the _x_ and _y_ positions based on
the cosinus and sinus of the angle.
Each feature has an angle (like in a clock). Starting from top (12h) as the Argumentation and continues clockwise until Aggressiveness.
NOTE: we need to multiply this value for the radius of the node and a little extra space for the dot radius.
If we do not add this little extra space, the dots will have its center position at the border of the node, what will cause overlapping
between dots for nodes with smaller radius.