Resize and reposition trivial cheese on the node + Solve bug
Solve bug: in function computeNodeRadius, we need to check if a value is null or undefined.
(For that we use *== null*)

Created functions **positionImage** and **sizeImage** to position an image associated with a node centered to the node
and resize it to be a *imgRatio* smaller than the node.
Set imgRatio = 20.

Modify function to draw features as trivial cheese on the node centered and resized.