Hide images for the root node + resize rectangular features
Change the way to check if a node is the root node (to avoid showing images about features or targets in it).

Create new function to draw rectangular features to call the functions positionImages and sizeImage with a bigger (20%) node radius
so the rectangle covers all the node (and we do not see the node circle under the rectangle).

ToDo: reduce coupling in the way we loop elements with lists of objects.