WIP: center and zoom to fit the initial visualization of the graph
Add functions **computeDimensions** and **zoomToFit**.
**computeDimensions** computes the box borders containing all the nodes.
WIP: **zoomToFit** should scale the visualization of the graph and center it in the canvas.

Update functions **zoom** and **zoomListener**.

BUG: console.log in **zoom** is not being printed.

NOTE: The initial visualization does not fit in the canvas.
After the first visualization, the zoom listener jumps to another part of the graph with a different scale.